Day6_Task/SVM_Kernel_Comparison/requirements.txt
numpy
pandas
matplotlib
seaborn 
mlflow
scikit-learn
python-dotenv
joblib